Termino dos desafios básicos

+++ Desafios de codigo/Soma de H com N Termos/src/Main.java
@@ -5,11 +5,11 @@ public class Main {
         double n = 4;
         for (int i = 1; i <= n; i++) {
             //TODO: Calcule o valor de H de forma que resulte na soma dos termos:
-            h = h + (1/i);
+            h = h + (1/(double)i);
         }
 
         //TODO: Imprima a soma dos termos, considerando um resultado com duas casas decimais:
         h = (int) Math.round(h);
-        System.out.println(h);
+        System.out.println((int)h);
     }
 }
